Refactor: Split TripPlanner.tsx into smaller components
This commit refactors the TripPlanner component into smaller, more manageable files to improve code organization and readability.
